backend/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dominate==2.9.1
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
gevent==23.9.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0
visitor==0.1.3
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
